requirements.txt
numpy==1.26.0
spacy==3.8.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.1/en_core_web_md-3.6.1-py3-none-any.whl
langdetect==1.0.9
requests
numpy
pandas
jsonschema
datetime
python-dotenv
prefect>=2.0